--- src/envs/.env
@@ -1,8 +1,0 @@
-PORT=4000
-BASE_URL=http://localhost:4000
-
-DATABASE_HOST=localhost
-DATABASE_NAME=postgres
-DATABASE_USER=postgres
-DATABASE_PASSWORD=mysecretpassword
-DATABASE_PORT=5432
