+++ src/shared/typeorm.service.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
+
 @Module({
   imports: [
     TypeOrmModule.forRootAsync({
@@ -14,7 +15,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
         username: config.get<string>('DATABASE_USER'),
         password: config.get<string>('DATABASE_PASSWORD'),
         entities: [__dirname + '/../**/*.entity.{js,ts}'],
-        synchronize: true, // never use TRUE in production!
+        // synchronize: true, // never use TRUE in production!
       }),
     }),
   ],
