+++ src/track/track.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from "@nestjs/common";
 import { TrackService } from './track.service';
 import { TrackController } from './track.controller';
-import { TrackEntity } from './entity/track.entity';
+import { TrackEntity } from './entities/track.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FavoritesModule } from '../favorites/favorites.module';
