+++ src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersEntity } from './entity/users.entity';
+import { UsersEntity } from './entities/users.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UsersEntity])],
